+++ packages/__tests__/src/icalendar/day-view.ts
@@ -3,7 +3,7 @@ import timeGridPlugin from '@fullcalendar/timegrid'
 import { EventSourceInput } from '@fullcalendar/core'
 import iCalendarPlugin from '@fullcalendar/icalendar'
 import oneHourMeeting from './data/oneHourMeeting'
-import recurringWeeklyMeeting from './data/recurringWeeklyMeeting'
+import recurringWeekly from './data/recurringWeekly'
 import mungedOneHourMeeting from './data/mungedOneHourMeeting'
 import meetingWithMungedStart from './data/meetingWithMungedStart'
 import alldayEvent from './data/alldayEvent'
@@ -35,9 +35,15 @@ describe('addICalEventSource with day view', () => {
   })
 
   it('adds a repeating weekly meeting', (done) => {
-    loadICalendarWith(recurringWeeklyMeeting, () => {
+    loadICalendarWith(recurringWeekly, () => {
       setTimeout(() => {
         assertEventCount(1)
+        const event = currentCalendar.getEvents()[0]
+        // test non-date props
+        expect(event.title).toBe('Weekly Monday meeting')
+        expect(event.url).toBe('https://fullcalendar.io/')
+        expect(event.extendedProps.description).toBe('this is the description')
+        expect(event.extendedProps.location).toBe('this is the location')
         done()
       }, 100)
     })
@@ -47,7 +53,13 @@ describe('addICalEventSource with day view', () => {
     loadICalendarWith(alldayEvent, () => {
       setTimeout(() => {
         assertEventCount(1)
-        currentCalendar.getEvents().forEach((event) => expect(event.allDay).toBeTruthy())
+        const events = currentCalendar.getEvents()
+        events.forEach((event) => expect(event.allDay).toBeTruthy())
+        // test non-date props
+        expect(events[0].title).toBe('First conference')
+        expect(events[0].url).toBe('https://fullcalendar.io/')
+        expect(events[0].extendedProps.description).toBe('this is the description')
+        expect(events[0].extendedProps.location).toBe('this is the location')
         done()
       })
     })
@@ -98,7 +110,7 @@ describe('addICalEventSource with day view', () => {
         setTimeout(() => {
           assertEventCount(1)
           const event = currentCalendar.getEvents()[0]
-          expect(event.end).toBeNull()
+          expect(event.end).toBe(null)
           done()
         }, 100)
       },