+++ packages/__tests__/src/icalendar/month-view.ts
@@ -8,7 +8,9 @@ import multidayEvent from './data/multidayEvent'
 import multipleMultidayEvents from './data/multipleMultidayEvents'
 import multipleEventsOneMunged from './data/multipleEventsOneMunged'
 import oneHourMeeting from './data/oneHourMeeting'
-import recurringWeeklyMeeting from './data/recurringWeeklyMeeting'
+import recurringWeekly from './data/recurringWeekly'
+import recurringWeeklyWithoutEnd from './data/recurringWeeklyWithoutEnd'
+import recurringWeeklyWithCount from './data/recurringWeeklyWithCount'
 import mungedOneHourMeeting from './data/mungedOneHourMeeting'
 
 describe('addICalEventSource with month view', () => {
@@ -26,8 +28,10 @@ describe('addICalEventSource with month view', () => {
   it('adds an all day event', (done) => {
     loadICalendarWith(alldayEvent, () => {
       setTimeout(() => {
+        let events = currentCalendar.getEvents()
+        expect(events[0].end).toBe(null)
+        events.forEach((event) => expect(event.allDay).toBeTruthy())
         assertEventCount(1)
-        currentCalendar.getEvents().forEach((event) => expect(event.allDay).toBeTruthy())
         done()
       }, 100)
     })
@@ -56,6 +60,9 @@ describe('addICalEventSource with month view', () => {
   it('adds a one-hour long meeting', (done) => {
     loadICalendarWith(oneHourMeeting, () => {
       setTimeout(() => {
+        let events = currentCalendar.getEvents()
+        expect(events[0].start).toEqualDate('2019-04-15T09:30:00')
+        expect(events[0].end).toEqualDate('2019-04-15T10:30:00')
         assertEventCount(1)
         currentCalendar.getEvents().forEach((event) => expect(event.allDay).not.toBeTruthy())
         done()
@@ -64,9 +71,36 @@ describe('addICalEventSource with month view', () => {
   })
 
   it('adds a repeating weekly meeting', (done) => {
-    loadICalendarWith(recurringWeeklyMeeting, () => {
+    loadICalendarWith(recurringWeekly, () => {
       setTimeout(() => {
+        let events = currentCalendar.getEvents()
+        expect(events[0].start).toEqualDate('2019-04-01T17:30:00')
+        expect(events[0].end).toEqualDate('2019-04-01T18:30:00')
         assertEventCount(6)
+        done()
+      }, 100)
+    })
+  })
+
+  it('adds a repeating weekly meeting, with null end', (done) => {
+    loadICalendarWith(recurringWeeklyWithoutEnd, () => {
+      setTimeout(() => {
+        let events = currentCalendar.getEvents()
+        expect(events[0].start).toEqualDate('2019-04-01T17:30:00')
+        expect(events[0].end).toBe(null)
+        assertEventCount(6)
+        done()
+      }, 100)
+    })
+  })
+
+  // https://github.com/fullcalendar/fullcalendar/issues/6190
+  // this feed starts at beginning of previous month (March 2019) and has 9 total occurences,
+  // 5 of which will be visible in the current month (April 2019)
+  it('adds a repeating weekly meeting, limited by COUNT, but across months', (done) => {
+    loadICalendarWith(recurringWeeklyWithCount, () => {
+      setTimeout(() => {
+        assertEventCount(5)
         done()
       }, 100)
     })
@@ -90,14 +124,14 @@ describe('addICalEventSource with month view', () => {
     })
   })
 
-  it('defaultAllDayEventDuration does not override ical default all day length of one day', (done) => {
+  it('defaultAllDayEventDuration overrides ical default all day length of one day', (done) => {
     loadICalendarWith(
       alldayEvent,
       () => {
         setTimeout(() => {
           assertEventCount(1)
           const event = currentCalendar.getEvents()[0]
-          expect(event.end.getDate()).toEqual(event.start.getDate() + 1)
+          expect(event.end.getDate()).toEqual(event.start.getDate() + 2)
           done()
         }, 100)
       },
@@ -110,12 +144,42 @@ describe('addICalEventSource with month view', () => {
     )
   })
 
+  it('calling refetchEvents request ical feed again', (done) => {
+    const feedUrl = '/mock.ics'
+    let fetchCnt = 0
+
+    XHRMock.get(feedUrl, (req, res) => {
+      fetchCnt += 1
+      return res.status(200)
+        .header('content-type', ICAL_MIME_TYPE)
+        .body(oneHourMeeting)
+    })
+
+    const calendar = initCalendar({
+      events: {
+        url: feedUrl,
+        format: 'ics',
+      },
+    })
+
+    setTimeout(() => {
+      expect(fetchCnt).toBe(1)
+      expect(calendar.getEvents().length).toBe(1)
+      calendar.refetchEvents()
+
+      setTimeout(() => {
+        expect(fetchCnt).toBe(2)
+        expect(calendar.getEvents().length).toBe(1)
+        done()
+      }, 100)
+    }, 100)
+  })
+
   function loadICalendarWith(rawICal: string, assertions: () => void, calendarSetup?: (source: EventSourceInput) => void) {
     const feedUrl = '/mock.ics'
 
     XHRMock.get(feedUrl, (req, res) => {
       expect(req.url().query).toEqual({})
-
       return res.status(200)
         .header('content-type', ICAL_MIME_TYPE)
         .body(rawICal)
@@ -134,6 +198,7 @@ describe('addICalEventSource with month view', () => {
 
   // Checks to make sure all events have been rendered and that the calendar
   // has internal info on all the events.
+  // TODO: don't use currentCalendar
   function assertEventCount(expectedCount: number) {
     expect(currentCalendar.getEvents().length).toEqual(expectedCount)